requirements.txt
# Dependências principais
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
plotly>=5.15.0
pyyaml>=6.0

# Algoritmos adicionais
xgboost>=1.7.0

# Utilitários
pathlib2>=2.3.7
python-dateutil>=2.8.0

# Para desenvolvimento e testes
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0

# Jupyter (opcional, para desenvolvimento)
jupyter>=1.0.0
ipykernel>=6.25.0